requirements.txt
attrs==23.2.0
Automat==22.10.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.51.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mpl-finance==0.10.1
mplfinance==0.12.10b0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2_binary==2.9.9
pyparsing==3.1.2
pyquery==2.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
cachelib==0.13.0
msgspec==0.18.6
Flask-Session==0.8.0
requests==2.31.0
six==1.16.0
TA-Lib-Precompiled==0.4.25
Twisted==24.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zope.interface==6.2
